Normalise file encodings
There are only two genuine bugfixes in this change:

- `HPM/events/Platine War.txt` which contained a general's name in a
  UTF-8 file
- `HPM/localisation/00_msc_localization.csv` which contained an initial
  BOM

Even this second one is arguable because the file contained no UTF-8
characters and Victoria 2 ignores the initial line of localisation files
anyway. That being said it was acting as a landmine of sorts, hinting
that this particular localisation file should be UTF-8 encoded--but this
is never correct.

The rest are encoding changes which only affect comments.