Add random variation to AI choice of NV
Before this change the AI only ever picked 'Order' as its national
value. This behaviour is rooted in how the game evaluates which
decisions to activate, proceeding from top to bottom & considering the
`ai_will_do` factor at each step---because that factor was 1, no other
decision were ever considered.

Setting this factor between 0 and 1 for the `switch_to_order` decision
ensures that other national values have a chance to be picked by the AI.
However this factor as well as the subsequent ones have to be computed
carefully. For the sake of future maintenance, the details of that
computation have been included in a comment beside the decisions.

This was favoured over e.g. using a `random_list` command as it was
found that multiple countries activating the decision on the same day
could all end up with the same choice. In practice this affected e.g.
Egypt and some of the Japanese countries, which are Westernised enough
at the start to activate the decision at the same time in the opening
days of the game.

-----

For the purpose of testing this change the following debug decision file
was used:

```
political_decisions = {
    reroll_nv = {
        potential = {}

        allow = {}

        effect = {
            any_country = {
                civilized = no
                nationalvalue = nv_tradition
                set_country_flag = nv_change
                clr_country_flag = switched_nv
                remove_country_modifier = national_confusion
            }
        }

        ai_will_do = { factor = 0 }
    }
}
```

When activated, the `reroll_nv` decision will ensure that all other
countries are eligible to (re)pick a national value the following day.
(Modulo some things like war and revolts.) Looking inside `game.log` for
entries of the kind "Decision: Embrace the Ideal of <national value> in
<country>" allows for tallying the probabilites of the AI picking each
national value.