Add steal_puppet to the peace order
CB `steal_puppet` was missing from the peace order list, which could
result in unintended results in a peace deal involving multiple CBs.

The CB is placed in-between the CBs that release a vassal and the CBs
that vassalise a country. This seems appropriate as in some sense
`steal_puppet` does those two things in succession.