Clean up sphere requirements on CBs
Various CBs had one of the two following sphere requirements:

- `part_of_sphere = no`
- `NOT = { in_sphere = THIS }`

As discussed in #47 the former is a legacy requirement. As to the latter
it is of little effect: the game already prevents fabrication or
declaration against one's own spherelings.

What practical effect these requirements did have (if any) was to
invalidate CBs handed out by event/decision, or that were active in
ongoing wars. (Though the latter has proved difficult to replicate
in-game.)

This change removes most (see next paragraph) such requirements to spare
frustration and confusion from the player. As a side-effect this
introduces some overlap in functionality between `add_to_sphere` and
`take_from_sphere`.

The purely sphere-making CBs (`add_to_sphere` and
`unification_add_to_sphere`) are changed to not work on one's own
sphereling specifically. Without this requirement a country could be
sitting on an `add_to_sphere` CB (when gained from an event/decision)
that would be displayed e.g. on the diplomacy screen but that they could
not use--unless they removed the country from their sphere, upon which
they could declare... to re-add the country to it.