Changed some Brazil trade goods to improve historical accuracy
Tobacco plantations in the middle of the jungle have been removed. Reduced spread of coffee plantations to the south. Increase of cattle in the interior of the country. Some minor climate corrections. Small increase in cotton, timber and rubber producing provinces.

Note about Brazil rubber exports: the first record is from 1832, when the United Kingdom imported 23 tons. Representing less than 1% of total Brazilian exports. Therefore, rubber market already existed in the game timeline, but with minimal production and low demand.