Condition adoption of writing system on modifier
The purpose of the `establish_writing_system` decision (typo in the name
fixed in this change) is to remove the crippling `lacks_writing_system`
modifier when applicable. However the decision is only conditioned on
the related `lacking_writing_system` flag and not the modifier itself.

Should a country lose the flag in some fashion, it would be stuck with
the debilitating modifier for the rest of the game. I don't believe that
this status can realistically be achieved in most games, though it is
not entirely outside the realm of possibility either. For instance an
organised `ARO` (aka Biafra) that managed impossibly high literacy
through colonial shenanigans might become eligible for `colonial_legacy`
after release.

With this change the condition for the `establish_writing_system`
decision correctly reflects its effect, namely by targeting the
`lacks_writing_system` modifier. On top of that the modifier is also
consistently removed whenever the flag is.