Gfx life_rating visual scale reduced by half and fixed vanilla needs_meter graphical glitch
To improve the visualization of life rating on the province's screen, the scale of the gfx was reduced by half, with 50 being the maximum.  (Using the same trick, it is also possible to show a scale of 0-10 life_rating, which would be better in my opinion, but it is necessary to test effects on migration and pop grow.)

A vanilla visual glitch in the needs meter was also fixed. The 'progress.fx' function cuts a line of pixels after processing the sprite, perhaps due to rounding error. Reduced and lightened beer foam, TGA transparency of the wine glass has been corrected.