Exclude uncivs from regencies
A large part of unciv gameplay is navigating your country's militancy
and consciousness. However with few options available to the player to
mitigate rises in discontentment, this more or less ends up as a race
for the country to achieve Westernisation before reactionary revolts are
no longer manageable.

In practical term this means that the regency event is a death sentence
to an unciv, unless the country was already close to 100%
Westernisation. This change limits the event to civilised countries
only.