Correct occurrences of formerly colonial MOS to BKF
At some point `MOS` was doing double duty as both a native and colonial
country. This was superseded by the use of `BKF` as a strictly colonial
country.

This change updates some old code to correctly refer to `BKF` cores
following organisation of the region. A review suggests that all
remaining `MOS`/`BKF` occurrences properly respect the native/colonial
separation.