Restrict Eiffel Tower World Fair exploit
France gets a vastly increased priority to be offered the chance to host
the next World Fair, but only so long as the Eiffel Tower hasn't been
built. This leads to an exploit where a crafty player never does build
it, to keep open the option to convert cash into prestige.

This change flips cause and effect around: France needs to have started
construction of the Eiffel Tower around, before being considered for
increased priority. Construction is still conditioned on
`bessemer_steel`, same as before.