Restore player/AI parity for colonial organization from Africa
For the player (but *not* the AI) it is not possible to access
organization decisions without first displaying the decision through
`turn_on_organize_colony_decisions`. Practically speaking, this means
that all de-cluttered organization decisions have their potential gated
by their own block *and* by that of the de-clutter decision.

As an example an African AI country who owns 1413 (Batavia) may
eventually satisfy enough conditions to activate `organize_indonesia`. A
player country in the same situation may not, because
`turn_on_organize_colony_decisions` also expects the capital to be out
of Africa.

This change modifies & relaxes the refactored checks on the de-clutter
decision. This prefix:

            NOT = {
                has_country_flag = organize_colony_decisions
                ai = yes
            }
            civilized = yes

now matches the following inside de-cluttered decisions:

            OR = {
                has_country_flag = organize_colony_decisions
                ai = yes
            }
            civilized = yes

Furthermore the overall capital check has been pushed out of the prefix,
down into specific clauses where applicable (e.g. when checking for the
ability to organize Nigeria). This is done in a way to consistently
mirror the actual potential block of each corresponding organization
decision: split diffing/editing can help.

For ***re***-organization decision, the capital check is added together
with a master (or similar flag) check. This is again done for
consistency.

Put all together & to demonstrate how each part mimics the other, first
the clauses for Nigeria:

    #Nigeria
    AND = {
        NOT = {
            has_global_flag = nigeria_organized
            capital_scope = { continent = africa }
        }
        OR = {
            owns = 1934
            owns = 1928
            owns = 1945
            owns = 1923
            owns = 1941
            owns = 1943
        }
    }
    AND = {
        has_global_flag = nigeria_organized
        owns = 1923
        owns = 1934
        owns = 1928
        NGR = {
            exists = no
            NOT = { is_culture_group = THIS }
        }
        NOT = {
            has_country_flag = nigeria_master
            capital_scope = { continent = africa }
        }
    }

Then the potential for `organize_nigeria`:

    OR = {
        has_country_flag = organize_colony_decisions
        ai = yes
    }
    civilized = yes
    NOT = {
        has_global_flag = nigeria_organized
        capital_scope = { continent = africa }
    }
    OR = {
        owns = 1934
        owns = 1928
        owns = 1945
        owns = 1923
        owns = 1941
        owns = 1943
    }

Lastly the potential for `reorganize_nigeria`:

    OR = {
        has_country_flag = organize_colony_decisions
        ai = yes
    }
    civilized = yes
    has_global_flag = nigeria_organized
    owns = 1923
    owns = 1934
    owns = 1928
    NGR = {
        exists = no
        NOT = { is_culture_group = THIS }
    }
    NOT = {
        has_country_flag = nigeria_master
        capital_scope = { continent = africa }
    }

Leaving no room for disparities between player and AI.