Add cost warnings to colonial organization decisions & event
Add missing cost/debt red warning to `EVTOPTD199041` when organising the
colonial RGOs of Sierra Leone.

The following organisation *decisions* (not events) have also been
extended with a similar warning. This is not the usual warning, since
the actual cost/debt is only incurred during the subsequent event and
the player may even pick a different option. They are however convenient
reminders, and follow in the footsteps of `[re]organize_nigeria` which
do just that.

- `[re]organize_sierra_leone`
- `[re]organize_sudan`
- `[re]organize_zambia`
- `[re]organize_mali`
- `[re]organize_indonesia`

Also bundled are the following unrelated editorial changes:

- capitalise the names of Botswanan peoples
- refer to Angola as *Angola* instead of *the Angola*, likewise
  Mozambique when referring to the country and not the oceanic feature